docs(tools): move k6 and add new structure
