+++ docs/docs/getting-started/try-live-demos.mdx
@@ -15,9 +15,9 @@ Tracetest has two shared demo environments which allow you to view, create, and 
 
 ## OpenTelemetry Astronomy Shop Demo
 
-The official [demo of the OpenTelemetry project](https://github.com/open-telemetry/opentelemetry-demo)!
-
 👉 **Join our [shared OpenTelemetry Demo environment](https://app.tracetest.io/organizations/ttorg_2179a9cd8ba8dfa5/invites/invite_646747b36e9ee487/accept).**
+
+The OpenTelemetry team maintains an official [demo of the OpenTelemetry project](https://github.com/open-telemetry/opentelemetry-demo).
 
 Once you join, check out the [OpenTelemetry Demo store docs](/live-examples/opentelemetry-store/overview/) to learn about:
 
@@ -28,9 +28,9 @@ Once you join, check out the [OpenTelemetry Demo store docs](/live-examples/open
 
 ## Pokeshop API Demo
 
-The Tracetest team maintains a fun Pokémon microservice app built using Node.js with two services, RabbitMQ communicating between them, and a Kafka stream.
-
 👉 **Join our [shared Pokeshop API Demo environment](https://app.tracetest.io/organizations/ttorg_2179a9cd8ba8dfa5/invites/invite_760904a64b4b9dc9/accept).**
+
+The Tracetest team maintains a fun Pokémon microservice app built using Node.js with two services, RabbitMQ communicating between them, and a Kafka stream.
 
 Once you join, check out the [Pokeshop API Demo docs](/live-examples/pokeshop/overview) to learn about:
