+++ docs/docs/tools-and-integrations/img/tracetest-cloud-k6.gif
