+++ examples/tracetest-core-k6/tests/test.yaml
@@ -1,0 +1,21 @@
+type: Test
+spec:
+  id: kc_MgKoVR
+  name: K6
+  description: K6
+  trigger:
+    type: traceid
+    traceid:
+      id: ${var:TRACE_ID}
+  specs:
+    - selector: span[tracetest.span.type="general" name="import pokemon"]
+      assertions:
+        - attr:tracetest.selected_spans.count = 1
+    - selector: span[tracetest.span.type="http" name="HTTP GET pokeapi.pokemon" http.method="GET"]
+      assertions:
+        - attr:http.url = "https://pokeapi.co/api/v2/pokemon/6"
+    - selector:
+        span[tracetest.span.type="database" name="create pokeshop.pokemon" db.system="postgres"
+        db.name="postgres" db.user="postgres" db.operation="create" db.sql.table="pokemon"]
+      assertions:
+        - attr:db.result | json_path '.name' = "charizard"
