+++ examples/tracetest-k6/.env.template
