+++ examples/tracetest-k6/assets/assertions.gif
