+++ examples/tracetest-k6/collector.config.yaml
@@ -8,24 +8,15 @@ processors:
   batch:
     timeout: 100ms
 
-  # Data sources: traces
-  probabilistic_sampler:
-    hash_seed: 22
-    sampling_percentage: 100
-
 exporters:
   jaeger:
     endpoint: jaeger:14250
     tls:
       insecure: true
-  otlp/2:
-    endpoint: tempo-us-central1.grafana.net:443
-    headers:
-      Authorization: Basic <your-auth-token>
 
 service:
   pipelines:
     traces:
       receivers: [otlp]
-      processors: [probabilistic_sampler, batch]
-      exporters: [otlp/2, jaeger]
+      processors: [batch]
+      exporters: [jaeger]
