+++ examples/tracetest-k6/tests/test.yaml
@@ -6,16 +6,18 @@ spec:
   trigger:
     type: traceid
     traceid:
-      id: ${var:TRACE_ID}
+      id: ${env:TRACE_ID}
   specs:
     - selector: span[tracetest.span.type="general" name="import pokemon"]
+      name: Should have imported the pokemon
       assertions:
         - attr:tracetest.selected_spans.count = 1
-    - selector: span[tracetest.span.type="http" name="HTTP GET pokeapi.pokemon" http.method="GET"]
+    - selector: |-
+        span[tracetest.span.type="http" net.peer.name="pokeapi.co" http.method="GET"]
+      name: Should trigger a request to the POKEAPI
       assertions:
-        - attr:http.url = "https://pokeapi.co/api/v2/pokemon/6"
-    - selector:
-        span[tracetest.span.type="database" name="create pokeshop.pokemon" db.system="postgres"
-        db.name="postgres" db.user="postgres" db.operation="create" db.sql.table="pokemon"]
+        - attr:http.url   =  "https://pokeapi.co/api/v2/pokemon/6"
+    - selector: span[tracetest.span.type="database" name="create postgres.pokemon"]
+      name: Should insert the pokemon to the DB
       assertions:
-        - attr:db.result | json_path '.name' = "charizard"
+        - attr:db.result | json_path '.name'  =  "charizard"
