+++ examples/tracetest-k6/tracetest-config.yaml
@@ -1,7 +1,7 @@
-postgres:
-  host: postgres
-  user: postgres
-  password: postgres
-  port: 5432
-  dbname: postgres
-  params: sslmode=disable
+# postgres:
+#   host: postgres
+#   user: postgres
+#   password: postgres
+#   port: 5432
+#   dbname: postgres
+#   params: sslmode=disable
